feat(ui): create side page
