+++ src/store/reducers/walletConnectionsCache/walletConnectionsCache.types.ts
@@ -1,6 +1,6 @@
 // TODO: mock data type for connect wallet ui. Need update after core function completed.
 interface ConnectionData {
-  id: number;
+  id: number | string;
   name: string;
   owner: string;
   image?: string;
