+++ src/ui/pages/SidePage/SidePage.types.ts
@@ -1,0 +1,6 @@
+interface SidePageContentProps {
+  open: boolean;
+  setOpenPage: (value: boolean) => void;
+}
+
+export type { SidePageContentProps };
