+++ src/ui/pages/SidePage/components/IncomingRequest/components/MultiSigRequest.scss
