+++ src/ui/pages/SidePage/components/IncomingRequest/components/SignTransactionRequest.scss
