+++ src/ui/pages/SidePage/components/WalletConnect/WalletConnect.scss
