+++ src/ui/pages/SidePage/components/WalletConnect/index.ts
