+++ src/ui/pages/SidePage/index.ts
@@ -1,0 +1,1 @@
+export * from "./SidePage";
